+++ packages/proto/src/messages/vesting/msgClawback.ts
@@ -1,0 +1,17 @@
+import { MsgClawback } from '../../types/evmos/vesting/tx'
+
+export function createMsgClawback(
+  funderAddress: string,
+  accountAddress: string,
+  destAddress?: string,
+) {
+  const msg = new MsgClawback({
+    funderAddress,
+    accountAddress,
+    destAddress,
+  })
+  return {
+    message: msg,
+    path: MsgClawback.typeName,
+  }
+}