+++ packages/proto/src/types/cosmos/crypto/keys/secp256k1/index.ts
@@ -1,0 +1,1 @@
+export * from '@buf/cosmos_cosmos-sdk.bufbuild_es/cosmos/crypto/secp256k1/keys_pb'