+++ packages/transactions/src/messages/vesting/msgClawback.ts
@@ -1,0 +1,78 @@
+import {
+  createMsgClawback as protoMsgClawback,
+  createTransaction,
+} from '@evmos/proto'
+
+import {
+  createEIP712,
+  generateFee,
+  generateMessage,
+  generateTypes,
+  createMsgClawback,
+  MSG_CLAWBACK,
+} from '@evmos/eip712'
+
+import { Chain, Fee, Sender } from '../common'
+
+export interface MessageMsgClawback {
+  funderAddress: string
+  accountAddress: string
+  destAddress?: string
+}
+
+export function createTxMsgClawback(
+  chain: Chain,
+  sender: Sender,
+  fee: Fee,
+  memo: string,
+  params: MessageMsgClawback,
+) {
+  // EIP712
+  const feeObject = generateFee(
+    fee.amount,
+    fee.denom,
+    fee.gas,
+    sender.accountAddress,
+  )
+  const types = generateTypes(MSG_CLAWBACK)
+
+  const msg = createMsgClawback(
+    params.funderAddress,
+    params.accountAddress,
+    params.destAddress,
+  )
+  const messages = generateMessage(
+    sender.accountNumber.toString(),
+    sender.sequence.toString(),
+    chain.cosmosChainId,
+    memo,
+    feeObject,
+    msg,
+  )
+  const eipToSign = createEIP712(types, chain.chainId, messages)
+
+  // Cosmos
+  const msgCosmos = protoMsgClawback(
+    params.funderAddress,
+    params.accountAddress,
+    params.destAddress,
+  )
+  const tx = createTransaction(
+    msgCosmos,
+    memo,
+    fee.amount,
+    fee.denom,
+    parseInt(fee.gas, 10),
+    'ethsecp256',
+    sender.pubkey,
+    sender.sequence,
+    sender.accountNumber,
+    chain.cosmosChainId,
+  )
+
+  return {
+    signDirect: tx.signDirect,
+    legacyAmino: tx.legacyAmino,
+    eipToSign,
+  }
+}