+++ packages/address-converter/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.9 (2022-11-14)
+
+### Bug Fixes
+
+- shx and link-module-alias change reverted ([#87](https://github.com/evmos/evmosjs/issues/87)) ([d4eeda5](https://github.com/evmos/evmosjs/commit/d4eeda59c8f5549d661e6bedd7a25df0b11afb9c))
+
+### Features
+
+- add akash support to recreate this pkg ([#89](https://github.com/evmos/evmosjs/issues/89)) ([5017449](https://github.com/evmos/evmosjs/commit/5017449d349b74be354f9d1615b0cef179cc13c8))
+- cosmos and osmosis address converter ([#19](https://github.com/evmos/evmosjs/issues/19)) ([96c0999](https://github.com/evmos/evmosjs/commit/96c0999b2b7b9c00a1be5b8d1f94587d5a22274a))
+- init repo ([#1](https://github.com/evmos/evmosjs/issues/1)) ([75a1c71](https://github.com/evmos/evmosjs/commit/75a1c71af1e48216139554f375151e167d3ff87f))
+- kyve address converter ([#39](https://github.com/evmos/evmosjs/issues/39)) ([20aad34](https://github.com/evmos/evmosjs/commit/20aad344eab10a0d3377341a7651881b9ec0b591))
+- proposal deposit message ([#86](https://github.com/evmos/evmosjs/issues/86)) ([5410c1f](https://github.com/evmos/evmosjs/commit/5410c1fe042c13a77beb5629a07371dd905559de))
+- proto objects for EIP712 ([#7](https://github.com/evmos/evmosjs/issues/7)) ([6ecd900](https://github.com/evmos/evmosjs/commit/6ecd9004f081c6a70b80d903878877d378ff6c75))
+
 ## [0.1.8](https://github.com/tharsis/evmosjs/compare/@evmos/address-converter@0.1.7...@evmos/address-converter@0.1.8) (2022-05-29)
 
 ### Features