+++ packages/proto/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.26](https://github.com/evmos/evmosjs/compare/@evmos/proto@0.1.25...@evmos/proto@0.1.26) (2022-11-14)
+
+### Bug Fixes
+
+- shx and link-module-alias change reverted ([#87](https://github.com/evmos/evmosjs/issues/87)) ([d4eeda5](https://github.com/evmos/evmosjs/commit/d4eeda59c8f5549d661e6bedd7a25df0b11afb9c))
+
+### Features
+
+- createValidator message ([#88](https://github.com/evmos/evmosjs/issues/88)) ([85df493](https://github.com/evmos/evmosjs/commit/85df493d5283b836e08aada2593f90c5fcfad516))
+
+## [0.1.25](https://github.com/evmos/evmosjs/compare/@evmos/proto@0.1.22...@evmos/proto@0.1.25) (2022-11-11)
+
+### Features
+
+- governance MsgSubmitProposal ([#80](https://github.com/evmos/evmosjs/issues/80)) ([419bae1](https://github.com/evmos/evmosjs/commit/419bae1f6dbf1e4d1a77d83f50a358f4fc91708a))
+- proposal deposit message ([#86](https://github.com/evmos/evmosjs/issues/86)) ([5410c1f](https://github.com/evmos/evmosjs/commit/5410c1fe042c13a77beb5629a07371dd905559de))
+- support for MsgSetWithdrawAddress ([#75](https://github.com/evmos/evmosjs/issues/75)) ([a879799](https://github.com/evmos/evmosjs/commit/a879799f06e57b2959b6a0c9bc622f540e3d8920))
+- vesting messages ([#85](https://github.com/evmos/evmosjs/issues/85)) ([ca75041](https://github.com/evmos/evmosjs/commit/ca750417ea7db5487de424d65d5d17517443ee01))
+
 ## [0.1.24](https://github.com/evmos/evmosjs/compare/@evmos/proto@0.1.22...@evmos/proto@0.1.24) (2022-10-10)
 
 ### Features